Comparative Model Upload
I am still working on a couple tweaks and messing around with different models for the data, that is why I am pushing this to a different branch for the time being. Working on cleaning it all up but want this as a history to go back to if need be.